refactor: shared server module (#1210)
