+++ electron/package.json
@@ -72,12 +72,12 @@
   "dependencies": {
     "@alumna/reflect": "^1.1.3",
     "@janhq/core": "link:./core",
+    "@janhq/server": "file:../server",
     "@npmcli/arborist": "^7.1.0",
     "@types/request": "^2.48.12",
     "@uiball/loaders": "^1.3.0",
     "electron-store": "^8.1.0",
     "electron-updater": "^6.1.7",
-    "fastify": "^4.25.0",
     "fs-extra": "^11.2.0",
     "node-fetch": "2",
     "pacote": "^17.0.4",
