+++ server/tsconfig.json
@@ -14,7 +14,8 @@
     "skipLibCheck": true,
     "paths": { "*": ["node_modules/*"] },
     "typeRoots": ["node_modules/@types"],
-    "ignoreDeprecations": "5.0"
+    "ignoreDeprecations": "5.0",
+    "declaration": true
   },
   // "sourceMap": true,
